Fix regression during CI testing of the SDK sysroot

See merge request steamrt/steam-runtime-tools!714

+++ subprojects/container-runtime/populate-depot.py
@@ -1357,7 +1357,7 @@ class Main:
         top: Path,
         writer: TextIO,
         *,
-        can_rename_files: bool = False,
+        minimize: bool = False,
         preserve_mode: bool = True,
         preserve_time: bool = True,
         skip_runtime_files: bool = False
@@ -1376,10 +1376,6 @@ class Main:
         sha256: Dict[Tuple[int, int], str] = {}
         # { [device, inode]: hashed name }
         hashed_names: Dict[Tuple[int, int], str] = {}
-
-        for name in unlink_later:
-            with suppress(FileNotFoundError):
-                (top / name).unlink()
 
         writer.write('#mtree\n')
         writer.write('. type=dir\n')
@@ -1460,7 +1456,7 @@ class Main:
                         short_hash = digest[:2] + '/' + digest[2:8]
                         fields.append(f'sha256={digest}')
 
-                        if can_rename_files and (
+                        if minimize and (
                             stat_info.st_nlink > 1
                             or name in differ_only_by_case
                             or name in not_windows_friendly
@@ -1491,14 +1487,17 @@ class Main:
                             rename[hashed_name] = name
                             unlink_later.add(name)
                         # else represent ./foo/bar as ./foo/bar
-                    else:
+                    elif minimize:
                         unlink_later.add(name)
 
                 elif stat.S_ISLNK(stat_info.st_mode):
                     fields.append('type=link')
                     fields.append(
                         f'link={self.octal_escape(os.readlink(member))}')
-                    unlink_later.add(name)
+
+                    if minimize:
+                        unlink_later.add(name)
+
                 elif stat.S_ISDIR(stat_info.st_mode):
                     fields.append('type=dir')
                 else:
@@ -1511,27 +1510,31 @@ class Main:
 
                 writer.write(' '.join(fields) + '\n')
 
-            if differ_only_by_case and not can_rename_files:
+            if differ_only_by_case and not minimize:
                 writer.write('\n')
                 writer.write('# Files whose names differ only by case:\n')
 
                 for name in sorted(differ_only_by_case):
                     writer.write('# {}\n'.format(self.octal_escape(name)))
 
-            if not_windows_friendly and not can_rename_files:
+            if not_windows_friendly and not minimize:
                 writer.write('\n')
                 writer.write('# Files whose names are not Windows-friendly:\n')
 
                 for name in sorted(not_windows_friendly):
                     writer.write('# {}\n'.format(self.octal_escape(name)))
 
-        for name, original in rename.items():
-            (top / name).parent.mkdir(parents=True, exist_ok=True)
-            (top / original).replace(top / name)
+        if minimize:
+            for name, original in rename.items():
+                (top / name).parent.mkdir(parents=True, exist_ok=True)
+                (top / original).replace(top / name)
 
-        for name in unlink_later:
-            with suppress(FileNotFoundError):
-                (top / name).unlink()
+            for name in unlink_later:
+                with suppress(FileNotFoundError):
+                    (top / name).unlink()
+        else:
+            assert not rename, rename
+            assert not unlink_later, unlink_later
 
         return lc_names
 
@@ -1542,7 +1545,7 @@ class Main:
             lc_names = self.write_mtree(
                 Path(self.depot),
                 writer,
-                can_rename_files=False,
+                minimize=False,
                 preserve_mode=False,
                 preserve_time=False,
                 skip_runtime_files=True,
@@ -1582,7 +1585,7 @@ class Main:
             lc_names = self.write_mtree(
                 Path(root) / 'files',
                 writer,
-                can_rename_files=True,
+                minimize=True,
             )
 
             if '.ref' not in lc_names:
